+++ apps/deploy-web/src/components/new-deployment/ManifestEdit.tsx
@@ -1,7 +1,6 @@
 "use client";
 import type { Dispatch, SetStateAction } from "react";
 import { useEffect, useMemo, useRef, useState } from "react";
-import { certificateManager } from "@akashnetwork/akashjs/build/certificates/certificate-manager";
 import { Alert, AlertDescription, Button, CustomTooltip, FileButton, Input, Snackbar, Spinner } from "@akashnetwork/ui/components";
 import { cn } from "@akashnetwork/ui/utils";
 import type { EncodeObject } from "@cosmjs/proto-signing";
@@ -16,13 +15,14 @@ import { useSnackbar } from "notistack";
 import { browserEnvConfig } from "@src/config/browser-env.config";
 import { useCertificate } from "@src/context/CertificateProvider";
 import { useSdlBuilder } from "@src/context/SdlBuilderProvider/SdlBuilderProvider";
+import { useServices } from "@src/context/ServicesProvider";
 import { useWallet } from "@src/context/WalletProvider";
 import { useImportSimpleSdl } from "@src/hooks/useImportSimpleSdl";
 import { useManagedWalletDenom } from "@src/hooks/useManagedWalletDenom";
 import { useWhen } from "@src/hooks/useWhen";
+import { useFeatureFlags } from "@src/queries/featureFlags";
 import { useDeploymentList } from "@src/queries/useDeploymentQuery";
 import { useDepositParams } from "@src/queries/useSaveSettings";
-import { analyticsService } from "@src/services/analytics/analytics.service";
 import sdlStore from "@src/store/sdlStore";
 import type { TemplateCreation } from "@src/types";
 import type { DepositParams } from "@src/types/deployment";
@@ -34,7 +34,6 @@ import { validateDeploymentData } from "@src/utils/deploymentUtils";
 import { Timer } from "@src/utils/timer";
 import { TransactionMessageData } from "@src/utils/TransactionMessageData";
 import { domainName, handleDocClick, UrlService } from "@src/utils/urlUtils";
-import { updateWallet } from "@src/utils/walletUtils";
 import { useSettings } from "../../context/SettingsProvider";
 import { DeploymentDepositModal } from "../deployments/DeploymentDepositModal";
 import { DeploymentMinimumEscrowAlertText } from "../sdl/DeploymentMinimumEscrowAlertText";
@@ -46,7 +45,7 @@ import ViewPanel from "../shared/ViewPanel";
 import type { SdlBuilderRefType } from "./SdlBuilder";
 import { SdlBuilder } from "./SdlBuilder";
 
-const TRIAL_DEPLOYMENT_LIMIT = 50;
+const TRIAL_DEPLOYMENT_LIMIT = 5;
 
 type Props = {
   onTemplateSelected: Dispatch<TemplateCreation | null>;
@@ -71,11 +70,14 @@ export const ManifestEdit: React.FunctionComponent<Props> = ({
   const [selectedSdlEditMode, setSelectedSdlEditMode] = useAtom(sdlStore.selectedSdlEditMode);
   const [isRepoInputValid, setIsRepoInputValid] = useState(false);
   const [sdlDenom, setSdlDenom] = useState("uakt");
+  const { data: features } = useFeatureFlags();
+  const isAnonymousFreeTrialEnabled = features?.allowAnonymousUserTrial;
 
+  const { analyticsService } = useServices();
   const { settings } = useSettings();
   const { address, signAndBroadcastTx, isManaged, isTrialing, isOnboarding } = useWallet();
   const router = useRouter();
-  const { loadValidCertificates, localCert, isLocalCertMatching, loadLocalCert, setSelectedCertificate } = useCertificate();
+  const { updateSelectedCertificate, genNewCertificateIfLocalIsInvalid } = useCertificate();
   const [, setDeploySdl] = useAtom(sdlStore.deploySdl);
   const muiTheme = useMuiTheme();
   const smallScreen = useMediaQuery(muiTheme.breakpoints.down("md"));
@@ -235,32 +237,23 @@ export const ManifestEdit: React.FunctionComponent<Props> = ({
         return;
       }
 
-      if (isTrialing && !isOnboarding) {
-        sdl = appendTrialAttribute(sdl, TRIAL_ATTRIBUTE);
-      } else if (isOnboarding) {
-        sdl = appendTrialAttribute(sdl, TRIAL_REGISTERED_ATTRIBUTE);
+      if (isAnonymousFreeTrialEnabled) {
+        if (isTrialing && !isOnboarding) {
+          sdl = appendTrialAttribute(sdl, TRIAL_ATTRIBUTE);
+        } else if (isOnboarding) {
+          sdl = appendTrialAttribute(sdl, TRIAL_REGISTERED_ATTRIBUTE);
+        }
       }
 
-      const dd = await createAndValidateDeploymentData(sdl, null, deposit, depositorAddress);
-
-      const validCertificates = await loadValidCertificates();
-      const currentCert = validCertificates.find(x => x.parsed === localCert?.certPem);
-      const isCertificateValidated = currentCert?.certificate?.state === "valid";
-      const isLocalCertificateValidated = !!localCert && isLocalCertMatching;
+      const [dd, newCert] = await Promise.all([createAndValidateDeploymentData(sdl, null, deposit, depositorAddress), genNewCertificateIfLocalIsInvalid()]);
 
       if (!dd) return;
 
       const messages: EncodeObject[] = [];
-      const hasValidCert = isCertificateValidated && isLocalCertificateValidated;
-      let _crtpem: string;
-      let _encryptedKey: string;
 
       // Create a cert if the user doesn't have one
-      if (!hasValidCert) {
-        const { cert: crtpem, publicKey: pubpem, privateKey: encryptedKey } = certificateManager.generatePEM(address);
-        _crtpem = crtpem;
-        _encryptedKey = encryptedKey;
-        messages.push(TransactionMessageData.getCreateCertificateMsg(address, crtpem, pubpem));
+      if (newCert) {
+        messages.push(TransactionMessageData.getCreateCertificateMsg(address, newCert.cert, newCert.publicKey));
       }
 
       messages.push(TransactionMessageData.getCreateDeploymentMsg(dd));
@@ -268,18 +261,8 @@ export const ManifestEdit: React.FunctionComponent<Props> = ({
 
       if (response) {
         // Set the new cert in storage
-        if (!hasValidCert) {
-          updateWallet(address, wallet => {
-            return {
-              ...wallet,
-              cert: _crtpem,
-              certKey: _encryptedKey
-            };
-          });
-          const validCerts = await loadValidCertificates();
-          loadLocalCert();
-          const currentCert = validCerts.find(x => x.parsed === _crtpem) || null;
-          setSelectedCertificate(currentCert);
+        if (newCert) {
+          await updateSelectedCertificate(newCert);
         }
 
         setDeploySdl(null);
@@ -320,8 +303,8 @@ export const ManifestEdit: React.FunctionComponent<Props> = ({
 
   const { data: deployments } = useDeploymentList(address);
   const trialDeploymentLimitReached = useMemo(() => {
-    return isTrialing && (deployments?.length || 0) >= TRIAL_DEPLOYMENT_LIMIT;
-  }, [deployments?.length, isTrialing]);
+    return isAnonymousFreeTrialEnabled && isTrialing && (deployments?.length || 0) >= TRIAL_DEPLOYMENT_LIMIT;
+  }, [deployments?.length, isTrialing, isAnonymousFreeTrialEnabled]);
 
   return (
     <>