+++ apps/deploy-web/src/pages/api/gitlab/refresh.ts
@@ -12,7 +12,7 @@ export default defineApiHandler({
   }),
   async handler({ body, res, services }) {
     const { NEXT_PUBLIC_GITLAB_CLIENT_ID, GITLAB_CLIENT_SECRET } = services.config;
-    const gitlabAuth = new GitlabAuth(NEXT_PUBLIC_GITLAB_CLIENT_ID as string, GITLAB_CLIENT_SECRET as string);
+    const gitlabAuth = new GitlabAuth(NEXT_PUBLIC_GITLAB_CLIENT_ID as string, GITLAB_CLIENT_SECRET as string, undefined, services.externalApiHttpClient);
 
     const tokens = await gitlabAuth.refreshTokensUsingRefreshToken(body.refreshToken);
     res.status(200).json(tokens);