+++ apps/deploy-web/src/queries/useMarketData.ts
@@ -7,10 +7,10 @@ import { ApiUrlService } from "@src/utils/apiUtils";
 import { QueryKeys } from "./queryKeys";
 
 export function useMarketData(options?: Omit<UseQueryOptions<MarketData, Error, any, QueryKey>, "queryKey" | "queryFn">) {
-  const { axios } = useServices();
+  const { publicConsoleApiHttpClient } = useServices();
   return useQuery<MarketData, Error>({
     queryKey: QueryKeys.getFinancialDataKey(),
-    queryFn: () => axios.get<MarketData>(ApiUrlService.marketData()).then(response => response.data),
+    queryFn: () => publicConsoleApiHttpClient.get<MarketData>(ApiUrlService.marketData()).then(response => response.data),
     ...options
   });
 }