+++ apps/deploy-web/src/services/app-di-container/app-di-container.ts
@@ -9,7 +9,7 @@ import {
   UsageHttpService,
   UserHttpService
 } from "@akashnetwork/http-sdk";
-import { StripeService } from "@akashnetwork/http-sdk/src/stripe/stripe.service";
+import { StripeService as HttpStripeService } from "@akashnetwork/http-sdk/src/stripe/stripe.service";
 import { LoggerService } from "@akashnetwork/logging";
 import { getTraceData } from "@sentry/nextjs";
 import { MutationCache, QueryCache, QueryClient } from "@tanstack/react-query";
@@ -24,6 +24,7 @@ import { createContainer } from "../container/createContainer";
 import { ErrorHandlerService } from "../error-handler/error-handler.service";
 import { ManagedWalletHttpService } from "../managed-wallet-http/managed-wallet-http.service";
 import { ProviderProxyService } from "../provider-proxy/provider-proxy.service";
+import { StripeService } from "../stripe/stripe.service";
 
 export const createAppRootContainer = (config: ServicesConfig) => {
   const apiConfig = { baseURL: config.BASE_API_MAINNET_URL };
@@ -56,16 +57,17 @@ export const createAppRootContainer = (config: ServicesConfig) => {
         ]
       }),
     stripe: () =>
-      container.applyAxiosInterceptors(new StripeService(apiConfig), {
+      container.applyAxiosInterceptors(new HttpStripeService(apiConfig), {
         request: [container.authService.withAnonymousUserHeader]
       }),
+    stripeService: () => new StripeService(),
     tx: () =>
       container.applyAxiosInterceptors(new TxHttpService(customRegistry, apiConfig), {
         request: [container.authService.withAnonymousUserHeader]
       }),
-    template: () => container.applyAxiosInterceptors(new TemplateHttpService(apiConfig), {}),
+    template: () => container.applyAxiosInterceptors(new TemplateHttpService(apiConfig)),
     usage: () =>
-      withInterceptors(new UsageHttpService(apiConfig), {
+      container.applyAxiosInterceptors(new UsageHttpService(apiConfig), {
         request: [container.authService.withAnonymousUserHeader]
       }),
     auth: () =>
@@ -81,7 +83,13 @@ export const createAppRootContainer = (config: ServicesConfig) => {
       container.applyAxiosInterceptors(new ApiKeyHttpService(), {
         request: [container.authService.withAnonymousUserHeader]
       }),
-    axios: () => container.createAxios(),
+    externalApiHttpClient: () =>
+      container.createAxios({
+        headers: {
+          "Content-Type": "application/json",
+          Accept: "application/json"
+        }
+      }),
     createAxios:
       () =>
       (options?: CreateAxiosDefaults): AxiosInstance =>
@@ -136,9 +144,9 @@ export interface ServicesConfig {
   apiUrlService: () => ApiUrlService;
 }
 
-function withInterceptors<T extends Axios | AxiosInstance = AxiosInstance>(axios: T, interceptors: Interceptors) {
-  interceptors.request?.forEach(interceptor => axios.interceptors.request.use(interceptor));
-  interceptors.response?.forEach(interceptor => axios.interceptors.response.use(interceptor));
+function withInterceptors<T extends Axios | AxiosInstance = AxiosInstance>(axios: T, interceptors?: Interceptors) {
+  interceptors?.request?.forEach(interceptor => axios.interceptors.request.use(interceptor));
+  interceptors?.response?.forEach(interceptor => axios.interceptors.response.use(interceptor));
   return axios;
 }